moved view files to src root
